Splitted components in different files
