+++ src/routes/SafeAppLandingPage/SafeAppLandingPage.tsx
@@ -2,30 +2,20 @@ import { ReactElement, useCallback, useEffect } from 'react'
 import { useSelector } from 'react-redux'
 import { useLocation } from 'react-router-dom'
 import styled from 'styled-components'
-import { Card, Title, Button, Text, Loader } from '@gnosis.pm/safe-react-components'
-import Divider from '@material-ui/core/Divider'
-import ClickAwayListener from '@material-ui/core/ClickAwayListener'
-import List from '@material-ui/core/List'
-import Popper from '@material-ui/core/Popper'
+import { Card, Title, Loader } from '@gnosis.pm/safe-react-components'
 
-import SuccessSvg from 'src/assets/icons/safe-created.svg'
-import DemoSvg from 'src/assets/icons/demo.svg'
-import { getChainById, isValidChainId } from 'src/config'
-import { demoSafeRoute, history, OPEN_SAFE_ROUTE, WELCOME_ROUTE } from 'src/routes/routes'
+import { isValidChainId } from 'src/config'
+import { history, WELCOME_ROUTE } from 'src/routes/routes'
 import { useAppList } from 'src/routes/safe/components/Apps/hooks/appList/useAppList'
 import { SafeApp } from 'src/routes/safe/components/Apps/types'
 import { getAppInfoFromUrl } from 'src/routes/safe/components/Apps/utils'
 import { userAccountSelector } from 'src/logic/wallets/store/selectors'
-import { useStateHandler } from 'src/logic/hooks/useStateHandler'
 import { setChainId } from 'src/logic/config/utils'
-import ProviderDisconnected from 'src/components/AppLayout/Header/components/ProviderInfo/ProviderDisconnected'
-import Provider from 'src/components/AppLayout/Header/components/Provider'
-import ConnectDetails from 'src/components/AppLayout/Header/components/ProviderDetails/ConnectDetails'
-import NetworkLabel from 'src/components/NetworkLabel/NetworkLabel'
-import Img from 'src/components/layout/Img'
-import Link from 'src/components/layout/Link'
-import { black300, grey400, secondary } from 'src/theme/variables'
 import useAsync from 'src/logic/hooks/useAsync'
+import SafeAppDetails from 'src/routes/SafeAppLandingPage/components/SafeAppsDetails'
+import CreateNewSafe from 'src/routes/SafeAppLandingPage/components/CreateNewSafe'
+import ConnectWallet from 'src/routes/SafeAppLandingPage/components/ConnectWallet'
+import TryDemoSafe from 'src/routes/SafeAppLandingPage/components/TryDemoSafe'
 
 function SafeAppLandingPage(): ReactElement {
   const { search } = useLocation()
@@ -81,7 +71,7 @@ function SafeAppLandingPage(): ReactElement {
     }
   }, [isLoading, safeAppDetails, isManifestError])
 
-  const availableChains = safeAppDetails?.chainIds
+  const availableChains = safeAppDetails?.chainIds || []
 
   const showLoader = isLoading || !safeAppDetails
 
@@ -122,104 +112,6 @@ function SafeAppLandingPage(): ReactElement {
 
 export default SafeAppLandingPage
 
-const SafeAppDetails = ({ iconUrl, name, description, availableChains }) => {
-  const showAvailableChains = availableChains?.length > 0
-
-  return (
-    <>
-      <DetailsContainer>
-        <SafeIcon src={iconUrl} />
-        <DescriptionContainer>
-          <SafeAppTitle size="sm">{name}</SafeAppTitle>
-          <div>{description}</div>
-        </DescriptionContainer>
-      </DetailsContainer>
-      <Separator />
-
-      {/* Available chains */}
-      {showAvailableChains && (
-        <>
-          <ChainLabel size="lg">Available networks</ChainLabel>
-          <ChainsContainer>
-            {availableChains.map((chainId) => (
-              <div key={chainId}>
-                <NetworkLabel networkInfo={getChainById(chainId)} />
-              </div>
-            ))}
-          </ChainsContainer>
-          <Separator />
-        </>
-      )}
-    </>
-  )
-}
-
-const CreateNewSafe = () => {
-  return (
-    <>
-      <BodyImage>
-        <Img alt="Vault" height={92} src={SuccessSvg} />
-      </BodyImage>
-
-      <Button size="lg" color="primary" variant="contained" component={Link} to={OPEN_SAFE_ROUTE}>
-        <Text size="xl" color="white">
-          Create new Safe
-        </Text>
-      </Button>
-    </>
-  )
-}
-
-const ConnectWallet = () => {
-  const { clickAway, open, toggle } = useStateHandler()
-
-  return (
-    <StyledProvider>
-      <Provider
-        info={<ProviderDisconnected />}
-        open={open}
-        toggle={toggle}
-        render={(providerRef) => (
-          <StyledPopper
-            anchorEl={providerRef.current}
-            open={open}
-            placement="bottom"
-            popperOptions={{ positionFixed: true }}
-          >
-            <ClickAwayListener onClickAway={clickAway} touchEvent={false}>
-              <List component="div">
-                <ConnectDetails />
-              </List>
-            </ClickAwayListener>
-          </StyledPopper>
-        )}
-      />
-    </StyledProvider>
-  )
-}
-
-const TryDemoSafe = ({ safeAppUrl }) => {
-  return (
-    <SafeDemoContainer>
-      <Title size="xs">Want to try the app before using it?</Title>
-
-      <BodyImage>
-        <Img alt="Demo" height={92} src={DemoSvg} />
-      </BodyImage>
-      {safeAppUrl && (
-        <StyledDemoButton
-          color="primary"
-          component={Link}
-          to={`${demoSafeRoute}?appUrl=${encodeURI(safeAppUrl)}`}
-          size="lg"
-          variant="outlined"
-        >
-          Try Demo
-        </StyledDemoButton>
-      )}
-    </SafeDemoContainer>
-  )
-}
 const Container = styled.main`
   display: flex;
   justify-content: center;
@@ -240,46 +132,6 @@ const LoaderContainer = styled.div`
   min-height: 400px;
 `
 
-const DetailsContainer = styled.div`
-  display: flex;
-`
-
-const SafeIcon = styled.img`
-  width: 90px;
-  height: 90px;
-`
-
-const SafeAppTitle = styled(Title)`
-  margin-top: 0px;
-  margin-bottom: 12px;
-`
-
-const DescriptionContainer = styled.div`
-  padding-left: 66px;
-  flex-grow: 1;
-`
-
-const Separator = styled(Divider)`
-  margin: 32px 0;
-`
-
-const ChainLabel = styled(Text)`
-  color: ${black300};
-`
-
-const ChainsContainer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-
-  && > div {
-    margin-top: 12px;
-    margin-right: 8px;
-  }
-`
-
-const StyledPopper = styled(Popper)`
-  z-index: 1301;
-`
 const ActionsContainer = styled.div`
   display: flex;
 `
@@ -287,24 +139,4 @@ const ActionsContainer = styled.div`
 const UserSafeContainer = styled.div`
   flex: 1 0 50%;
   text-align: center;
-`
-
-const SafeDemoContainer = styled.div`
-  flex: 1 0 50%;
-  text-align: center;
-`
-const StyledProvider = styled.div`
-  width: 300px;
-  height: 56px;
-  margin: 0 auto;
-  border-radius: 8px;
-  border: 2px solid ${grey400};
-`
-
-const BodyImage = styled.div`
-  margin: 30px 0;
-`
-
-const StyledDemoButton = styled(Button)`
-  border: 2px solid ${secondary};
 `
